Add colors for robot
This commit tries to add color to my_robot. This is done by
both update the URDF file to refrect the real color in rViz,
as well as update the gazebo color.

This also removes any references to the robot in my_world,
which made the color disappear, as well as add a simple script
to spawn the robot in the middle.